requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.7.4.post0
aiosignal==1.3.1
anyio==4.3.0
APScheduler==3.10.1
async-timeout==3.0.1
attrs==24.2.0
cachetools==4.2.2
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.4.0
croniter==5.0.1
discord-components==0.0.0.1
discord-py-interactions==5.13.2
discord-typings==0.9.0
discord-ui==5.1.6
discord.py==2.4.0
emoji==2.14.0
frozenlist==1.5.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
idna==3.6
multidict==6.1.0
oauth2client==4.1.3
oauthlib==3.3.1
propcache==0.2.0
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-telegram-bot==22.3
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
Telethon==1.40.0
tomli==2.1.0
tornado==6.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.6
yarl==1.17.1
